add change date function

+++ ToDoList/data/functions.php
@@ -1,6 +1,8 @@
 <?php
 include_once "database.php";
 
+	// Form request, add list
+
 if (isset($_POST['name'])){
 		// validate input
 	$name = mysqli_real_escape_string($con, strip_tags($_POST['name']));
@@ -22,31 +24,64 @@ if (isset($_POST['name'])){
 			}
 		}
 }
-if (isset($_POST['id'])){
+
+	// Delete list request
+
+if (isset($_POST['del']) && isset($_POST['id'])){
 	$id = $_POST['id'];
 	$sql = "DELETE FROM list WHERE id = '$id'";
 	$result = mysqli_query($con, $sql);
 	if(!$result){
 		echo "Error: ".mysqli_error($con) ;
 	} else{
-		// header("Location: ../index.php");
-		// exit();
+		echo "Item deleted";
 	}
 }
 
-// function addList($con, $data) {
-// 	$name = mysqli_real_escape_string($con, strip_tags($data));
-// 	$start_date = strftime('%F');
-// 	$sql = "INSERT INTO list(name, start_date) VALUE('$name', '$start_date')";
-// 	$result = mysqli_query($con, $sql);
-// 	if (!$result) {
-// 		$result_error = mysqli_error($con);
-// 		echo $result_error;
-// 		return false;
-// 	}
-// 	// header("Location: index.php");
-// 	return true;
-// }
+if (isset($_POST['change']) && isset($_POST['id'])){
+	changeDoneDate($con, $_POST['id']);
+	// $result = mysqli_query($con, $sql);
+	// if(!$result){
+	// 	echo "Error: ".mysqli_error($con) ;
+	// } else{
+	// 	echo "Item deleted";
+	// }
+}
+
+function selectById($con, $id) {
+	$sql = "SELECT id, name, start_date, done FROM list WHERE id = '$id' ";
+	$result = mysqli_query($con, $sql);
+	if (!$result) {
+		$result_error = mysqli_error($con);
+		echo $result_error;
+		return false;
+	} else {
+		$row = mysqli_fetch_assoc($result);
+		return $row;
+	}
+}
+function updateDate($con, $done, $id){
+	$sql = "UPDATE list SET done = '$done' WHERE id = '$id' ";
+	$result = mysqli_query($con, $sql);
+}
+function changeDoneDate($con, $id){
+	if(!($row = selectById($con, $id))){
+		echo "Error: can't change list";
+	} else{
+		if (!$row['done']) {
+			$done = strftime('%F');
+			updateDate($con, $done, $id);
+
+		} else {
+			$done = null;
+			updateDate($con, $done, $id);
+		}
+		echo($done);
+	}
+
+}
+
+	// Show To-Do list function
 
 function getToDO($con) {
 	$sql = "SELECT id, name, start_date, done FROM list ORDER BY id DESC";
